ScalaTest in Docker at Travis CI.
This is a simple Hello World with respective test.

Issuing 'make docker' will build a Centos 7 Docker image with
git, scala, and sbt, which will be used for testing.

To run the tests within the Docker image type 'make docker-test'
after building the image.

This is used to run tests at Travis. Here we are taking advantage of
Travis' cache, so that dependencies are only downloaded once. Because
the environment variable $HOME is used in the Makefile this works both
on Travis and on your computer.

Using project dir contents instead of cloning the master branch.

If the master branch was used we couldn't test if a branch was causing
compilation errors with Travis. The master would always be used instead.

Added missing newlines and a help target to the makefile.

Now using docker container with scala instead of installing from source.

Edited Travis' email notification settings.

Added Slack integration to .travis.yml